services/currency-service/requirements.txt
Django==5.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
requests==2.31.0
python-decouple==3.8
